add throttle control
