+++ requirements.txt
@@ -1,48 +1,66 @@
+aioredis==2.0.1
 anyio==3.6.1
 asgiref==3.5.2
 asttokens==2.0.5
+async-timeout==4.0.2
+attrs==21.4.0
 autopep8==1.6.0
 certifi==2022.5.18.1
 charset-normalizer==2.0.12
 click==8.1.3
 coinbase==2.1.0
 colorama==0.4.4
+Deprecated==1.2.13
 dnspython==2.2.1
 email-validator==1.2.1
 executing==0.8.3
 fastapi==0.78.0
+fastapi-limiter==0.1.4
+flake8==4.0.1
 Flask==2.1.2
 h11==0.13.0
+hiredis==2.0.0
 httptools==0.4.0
 icecream==2.1.2
 idna==3.3
+iniconfig==1.1.1
 itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.1
+mccabe==0.6.1
 nulltype==2.3.1
-numpy==1.22.4
+numpy==1.23.1
 orjson==3.6.8
-pandas==1.4.2
-plaid-python==9.4.0
+packaging==21.3
+pandas==1.4.3
+plaid-python==9.7.0
+pluggy==1.0.0
+py==1.11.0
 pycodestyle==2.8.0
 pycryptodome==3.14.1
 pydantic==1.9.1
+pyflakes==2.4.0
 Pygments==2.12.0
+pyparsing==3.0.9
+pytest==7.1.2
 python-dateutil==2.8.2
 python-dotenv==0.20.0
 python-multipart==0.0.5
 pytz==2022.1
 PyYAML==6.0
-requests==2.27.1
+redis==4.3.4
+requests==2.28.1
 six==1.16.0
 sniffio==1.2.0
 starlette==0.19.1
 toml==0.10.2
+tomli==2.0.1
 typing_extensions==4.2.0
 ujson==5.3.0
-urllib3==1.26.9
-uvicorn==0.17.6
+urllib3==1.26.10
+uvicorn==0.18.2
 uvloop==0.16.0
 watchgod==0.8.2
 websockets==10.3
 Werkzeug==2.1.2
+wrapt==1.14.1
